Moved harmonic_subset into harmonic.py and combined the test files of both modules. Also fixed an issue where a grpah passed to harmonic.py could be reversed during the execution of the function.
